Revert "Process powerbalance parameters (#848)"
This reverts commit 3ce817315f64e0835dbcac621aa77790ae731872.
