+++ bluemira/codes/plasmod/mapping.py
@@ -22,8 +22,24 @@
 """
 PLASMOD mappings
 """
+from enum import Enum
 
-from bluemira.codes.utilities import Model, create_mapping
+from bluemira.base.look_and_feel import bluemira_print
+from bluemira.codes.utilities import create_mapping
+
+
+class Model(Enum):
+    """
+    Base Model Enum
+    """
+
+    @classmethod
+    def info(cls):
+        """
+        Show Model options
+        """
+        infostr = f"{cls.__doc__}\n" + "\n".join(repr(l_) for l_ in list(cls))
+        bluemira_print(infostr)
 
 
 class ImpurityModel(Model):
