+++ data/reactors/EU-DEMO/systems_code/mockPROCESS.json
@@ -54,12 +54,5 @@
     "tk_tf_wp": 0.70786,
     "tk_ts": 0.05,
     "tk_vv_in": 0.3,
-    "v_burn": 0.044484,
-    "TF_res_bus": 0,
-    "TF_res_tot": 0,
-    "TF_E_stored": 0,
-    "TF_respc_ob": 0,
-    "TF_currpt_ob": 0,
-    "P_bd_in": 0,
-    "condrad_cryo_heat": 0
+    "v_burn": 0.044484
 }
